Refine `count_nonzero` typing stub

+++ numpy/typing/tests/data/reveal/numeric.pyi
@@ -25,9 +25,9 @@ AR_O: npt.NDArray[np.object_]
 B: list[int]
 C: SubClass
 
-assert_type(np.count_nonzero(i8), int)
-assert_type(np.count_nonzero(AR_i8), int)
-assert_type(np.count_nonzero(B), int)
+assert_type(np.count_nonzero(i8), np.intp)
+assert_type(np.count_nonzero(AR_i8), np.intp)
+assert_type(np.count_nonzero(B), np.intp)
 assert_type(np.count_nonzero(AR_i8, keepdims=True), npt.NDArray[np.intp])
 assert_type(np.count_nonzero(AR_i8, axis=0), Any)
